Ya funciona todo canchas de nuevo
se cae de nuevo y no se que mas hacer

+++ ProyectoDeportivoCR/Repositories/ICanchaRepository.cs
@@ -4,8 +4,8 @@
     {
         Task<HttpResponseMessage> RegistrarCancha(CanchaModel model, string? token);
         Task<HttpResponseMessage> ActualizarInformacionCancha(CanchaModel model, string? token);
-        Task<HttpResponseMessage> DeshabilitarCancha(int canchaId, string? token);
-        Task<HttpResponseMessage> ObtenerCancha(int canchaId, string? token);
+        Task<HttpResponseMessage> DeshabilitarCancha(long canchaId, string? token);
+        Task<HttpResponseMessage> ObtenerCancha(long canchaId, string? token);
         Task<HttpResponseMessage> ObtenerTodasLasCanchas(string? token);
     }
 }
